types: remove unused solana_account_meta type
This type doesn't require a serializer, move it to solcap
