+++ src/flamenco/types/fd_types.c
@@ -918,135 +918,6 @@ void fd_solana_account_stored_meta_walk( void * w, fd_solana_account_stored_meta
   fun( w, self->pubkey, "pubkey", FD_FLAMENCO_TYPE_HASH256, "uchar[32]", level, 0  );
   fun( w, self, name, FD_FLAMENCO_TYPE_MAP_END, "fd_solana_account_stored_meta", level--, 0 );
 }
-int fd_solana_account_meta_encode( fd_solana_account_meta_t const * self, fd_bincode_encode_ctx_t * ctx ) {
-  int err;
-  err = fd_bincode_uint64_encode( self->lamports, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_uint64_encode( self->rent_epoch, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bytes_encode( self->owner, sizeof(self->owner), ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bool_encode( (uchar)(self->executable), ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bytes_encode( self->padding, 3, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  return FD_BINCODE_SUCCESS;
-}
-static int fd_solana_account_meta_decode_footprint_inner( fd_bincode_decode_ctx_t * ctx, ulong * total_sz ) {
-  if( ctx->data>=ctx->dataend ) { return FD_BINCODE_ERR_OVERFLOW; };
-  int err = 0;
-  err = fd_bincode_uint64_decode_footprint( ctx );
-  if( FD_UNLIKELY( err!=FD_BINCODE_SUCCESS ) ) return err;
-  err = fd_bincode_uint64_decode_footprint( ctx );
-  if( FD_UNLIKELY( err!=FD_BINCODE_SUCCESS ) ) return err;
-  err = fd_bincode_bytes_decode_footprint( 32, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bool_decode_footprint( ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bytes_decode_footprint( 3, ctx );
-  if( FD_UNLIKELY( err!=FD_BINCODE_SUCCESS ) ) return err;
-  return 0;
-}
-int fd_solana_account_meta_decode_footprint( fd_bincode_decode_ctx_t * ctx, ulong * total_sz ) {
-  *total_sz += sizeof(fd_solana_account_meta_t);
-  void const * start_data = ctx->data;
-  int err = fd_solana_account_meta_decode_footprint_inner( ctx, total_sz );
-  if( ctx->data>ctx->dataend ) { return FD_BINCODE_ERR_OVERFLOW; };
-  ctx->data = start_data;
-  return err;
-}
-static void fd_solana_account_meta_decode_inner( void * struct_mem, void * * alloc_mem, fd_bincode_decode_ctx_t * ctx ) {
-  fd_solana_account_meta_t * self = (fd_solana_account_meta_t *)struct_mem;
-  fd_bincode_uint64_decode_unsafe( &self->lamports, ctx );
-  fd_bincode_uint64_decode_unsafe( &self->rent_epoch, ctx );
-  fd_bincode_bytes_decode_unsafe( &self->owner[0], sizeof(self->owner), ctx );
-  fd_bincode_bool_decode_unsafe( &self->executable, ctx );
-  fd_bincode_bytes_decode_unsafe( self->padding, 3, ctx );
-}
-void * fd_solana_account_meta_decode( void * mem, fd_bincode_decode_ctx_t * ctx ) {
-  fd_solana_account_meta_t * self = (fd_solana_account_meta_t *)mem;
-  fd_solana_account_meta_new( self );
-  void * alloc_region = (uchar *)mem + sizeof(fd_solana_account_meta_t);
-  void * * alloc_mem = &alloc_region;
-  fd_solana_account_meta_decode_inner( mem, alloc_mem, ctx );
-  return self;
-}
-void fd_solana_account_meta_new(fd_solana_account_meta_t * self) {
-  fd_memset( self, 0, sizeof(fd_solana_account_meta_t) );
-}
-void fd_solana_account_meta_walk( void * w, fd_solana_account_meta_t const * self, fd_types_walk_fn_t fun, const char *name, uint level, uint varint ) {
-  (void) varint;
-  fun( w, self, name, FD_FLAMENCO_TYPE_MAP, "fd_solana_account_meta", level++, 0 );
-  fun( w, &self->lamports, "lamports", FD_FLAMENCO_TYPE_ULONG, "ulong", level, 0  );
-  fun( w, &self->rent_epoch, "rent_epoch", FD_FLAMENCO_TYPE_ULONG, "ulong", level, 0  );
-  fun( w, self->owner, "owner", FD_FLAMENCO_TYPE_HASH256, "uchar[32]", level, 0  );
-  fun( w, &self->executable, "executable", FD_FLAMENCO_TYPE_BOOL, "bool", level, 0  );
-  fun(w, self->padding, "padding", FD_FLAMENCO_TYPE_UCHAR, "uchar", level, 0 );
-  fun( w, self, name, FD_FLAMENCO_TYPE_MAP_END, "fd_solana_account_meta", level--, 0 );
-}
-int fd_solana_account_hdr_encode( fd_solana_account_hdr_t const * self, fd_bincode_encode_ctx_t * ctx ) {
-  int err;
-  err = fd_solana_account_stored_meta_encode( &self->meta, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_solana_account_meta_encode( &self->info, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bytes_encode( self->padding, 4, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_hash_encode( &self->hash, ctx );
-  if( FD_UNLIKELY( err ) ) return err;
-  return FD_BINCODE_SUCCESS;
-}
-static int fd_solana_account_hdr_decode_footprint_inner( fd_bincode_decode_ctx_t * ctx, ulong * total_sz ) {
-  if( ctx->data>=ctx->dataend ) { return FD_BINCODE_ERR_OVERFLOW; };
-  int err = 0;
-  err = fd_solana_account_stored_meta_decode_footprint_inner( ctx, total_sz );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_solana_account_meta_decode_footprint_inner( ctx, total_sz );
-  if( FD_UNLIKELY( err ) ) return err;
-  err = fd_bincode_bytes_decode_footprint( 4, ctx );
-  if( FD_UNLIKELY( err!=FD_BINCODE_SUCCESS ) ) return err;
-  err = fd_hash_decode_footprint_inner( ctx, total_sz );
-  if( FD_UNLIKELY( err ) ) return err;
-  return 0;
-}
-int fd_solana_account_hdr_decode_footprint( fd_bincode_decode_ctx_t * ctx, ulong * total_sz ) {
-  *total_sz += sizeof(fd_solana_account_hdr_t);
-  void const * start_data = ctx->data;
-  int err = fd_solana_account_hdr_decode_footprint_inner( ctx, total_sz );
-  if( ctx->data>ctx->dataend ) { return FD_BINCODE_ERR_OVERFLOW; };
-  ctx->data = start_data;
-  return err;
-}
-static void fd_solana_account_hdr_decode_inner( void * struct_mem, void * * alloc_mem, fd_bincode_decode_ctx_t * ctx ) {
-  fd_solana_account_hdr_t * self = (fd_solana_account_hdr_t *)struct_mem;
-  fd_solana_account_stored_meta_decode_inner( &self->meta, alloc_mem, ctx );
-  fd_solana_account_meta_decode_inner( &self->info, alloc_mem, ctx );
-  fd_bincode_bytes_decode_unsafe( self->padding, 4, ctx );
-  fd_hash_decode_inner( &self->hash, alloc_mem, ctx );
-}
-void * fd_solana_account_hdr_decode( void * mem, fd_bincode_decode_ctx_t * ctx ) {
-  fd_solana_account_hdr_t * self = (fd_solana_account_hdr_t *)mem;
-  fd_solana_account_hdr_new( self );
-  void * alloc_region = (uchar *)mem + sizeof(fd_solana_account_hdr_t);
-  void * * alloc_mem = &alloc_region;
-  fd_solana_account_hdr_decode_inner( mem, alloc_mem, ctx );
-  return self;
-}
-void fd_solana_account_hdr_new(fd_solana_account_hdr_t * self) {
-  fd_memset( self, 0, sizeof(fd_solana_account_hdr_t) );
-  fd_solana_account_stored_meta_new( &self->meta );
-  fd_solana_account_meta_new( &self->info );
-  fd_hash_new( &self->hash );
-}
-void fd_solana_account_hdr_walk( void * w, fd_solana_account_hdr_t const * self, fd_types_walk_fn_t fun, const char *name, uint level, uint varint ) {
-  (void) varint;
-  fun( w, self, name, FD_FLAMENCO_TYPE_MAP, "fd_solana_account_hdr", level++, 0 );
-  fd_solana_account_stored_meta_walk( w, &self->meta, fun, "meta", level, 0 );
-  fd_solana_account_meta_walk( w, &self->info, fun, "info", level, 0 );
-  fun(w, self->padding, "padding", FD_FLAMENCO_TYPE_UCHAR, "uchar", level, 0 );
-  fd_hash_walk( w, &self->hash, fun, "hash", level, 0 );
-  fun( w, self, name, FD_FLAMENCO_TYPE_MAP_END, "fd_solana_account_hdr", level--, 0 );
-}
 int fd_delegation_encode( fd_delegation_t const * self, fd_bincode_encode_ctx_t * ctx ) {
   int err;
   err = fd_pubkey_encode( &self->voter_pubkey, ctx );
