+++ src/flamenco/types/fd_types.h
@@ -176,27 +176,6 @@ struct __attribute__((packed)) fd_solana_account_stored_meta {
 typedef struct fd_solana_account_stored_meta fd_solana_account_stored_meta_t;
 #define FD_SOLANA_ACCOUNT_STORED_META_ALIGN (8UL)
 
-/* Encoded Size: Fixed (52 bytes) */
-struct __attribute__((packed)) fd_solana_account_meta {
-  ulong lamports;
-  ulong rent_epoch;
-  uchar owner[32];
-  uchar executable;
-  uchar padding[3];
-};
-typedef struct fd_solana_account_meta fd_solana_account_meta_t;
-#define FD_SOLANA_ACCOUNT_META_ALIGN (8UL)
-
-/* Encoded Size: Fixed (136 bytes) */
-struct __attribute__((packed)) fd_solana_account_hdr {
-  fd_solana_account_stored_meta_t meta;
-  fd_solana_account_meta_t info;
-  uchar padding[4];
-  fd_hash_t hash;
-};
-typedef struct fd_solana_account_hdr fd_solana_account_hdr_t;
-#define FD_SOLANA_ACCOUNT_HDR_ALIGN (8UL)
-
 /* https://github.com/solana-labs/solana/blob/8f2c8b8388a495d2728909e30460aa40dcc5d733/sdk/program/src/stake/state.rs#L303 */
 /* Encoded Size: Fixed (64 bytes) */
 struct fd_delegation {
@@ -2065,22 +2044,6 @@ static inline int fd_solana_account_stored_meta_decode_footprint( fd_bincode_dec
   return 0;
 }
 void * fd_solana_account_stored_meta_decode( void * mem, fd_bincode_decode_ctx_t * ctx );
-
-void fd_solana_account_meta_new( fd_solana_account_meta_t * self );
-int fd_solana_account_meta_encode( fd_solana_account_meta_t const * self, fd_bincode_encode_ctx_t * ctx );
-void fd_solana_account_meta_walk( void * w, fd_solana_account_meta_t const * self, fd_types_walk_fn_t fun, const char *name, uint level, uint varint );
-static inline ulong fd_solana_account_meta_size( fd_solana_account_meta_t const * self ) { (void)self; return 52UL; }
-static inline ulong fd_solana_account_meta_align( void ) { return FD_SOLANA_ACCOUNT_META_ALIGN; }
-int fd_solana_account_meta_decode_footprint( fd_bincode_decode_ctx_t * ctx, ulong * total_sz );
-void * fd_solana_account_meta_decode( void * mem, fd_bincode_decode_ctx_t * ctx );
-
-void fd_solana_account_hdr_new( fd_solana_account_hdr_t * self );
-int fd_solana_account_hdr_encode( fd_solana_account_hdr_t const * self, fd_bincode_encode_ctx_t * ctx );
-void fd_solana_account_hdr_walk( void * w, fd_solana_account_hdr_t const * self, fd_types_walk_fn_t fun, const char *name, uint level, uint varint );
-static inline ulong fd_solana_account_hdr_size( fd_solana_account_hdr_t const * self ) { (void)self; return 136UL; }
-static inline ulong fd_solana_account_hdr_align( void ) { return FD_SOLANA_ACCOUNT_HDR_ALIGN; }
-int fd_solana_account_hdr_decode_footprint( fd_bincode_decode_ctx_t * ctx, ulong * total_sz );
-void * fd_solana_account_hdr_decode( void * mem, fd_bincode_decode_ctx_t * ctx );
 
 static inline void fd_delegation_new( fd_delegation_t * self ) { fd_memset( self, 0, sizeof(fd_delegation_t) ); }
 int fd_delegation_encode( fd_delegation_t const * self, fd_bincode_encode_ctx_t * ctx );
