+++ src/flamenco/types/fd_types.json
@@ -168,29 +168,6 @@
       ]
     },
     {
-      "name": "solana_account_meta",
-      "packed": true,
-      "type": "struct",
-      "fields": [
-        { "name": "lamports", "type": "ulong" },
-        { "name": "rent_epoch", "type": "ulong" },
-        { "name": "owner", "type": "uchar[32]" },
-        { "name": "executable", "type": "bool" },
-        { "name": "padding", "type": "array", "element": "uchar", "length": 3 }
-      ]
-    },
-    {
-      "name": "solana_account_hdr",
-      "packed": true,
-      "type": "struct",
-      "fields": [
-        { "name": "meta", "type": "solana_account_stored_meta" },
-        { "name": "info", "type": "solana_account_meta" },
-        { "name": "padding", "type": "array", "element": "uchar", "length": 4 },
-        { "name": "hash", "type": "hash" }
-      ]
-    },
-    {
       "name": "delegation",
       "type": "struct",
       "fields": [
